+++ src/studio/src/designer/backend/Views/ServiceDevelopment/index.cshtml
@@ -24,12 +24,12 @@
 
 <!-- These imports are imported multiple times, we will need to clean up afterwards -->
 <link rel="stylesheet" href="~/designer/css/bootstrap.min.css">
-<link rel="stylesheet" href="~/designer/css/react/app-development.css">
+<link rel="stylesheet" href="http://localhost:8080/app-development.css">
 
 <div id="root" class="media-body flex-column d-flex"></div>
-<script src="~/designer/js/react/app-development.js" asp-append-version="true" type="text/javascript"></script>
-<script src="~/designer/js/react/1.app-development.js" asp-append-version="true" type="text/javascript"></script>
-<script src="~/designer/js/react/2.app-development.js" asp-append-version="true" type="text/javascript"></script>
-<script src="~/designer/js/react/3.app-development.js" asp-append-version="true" type="text/javascript"></script>
-<script src="~/designer/js/react/4.app-development.js" asp-append-version="true" type="text/javascript"></script>
+<script src="http://localhost:8080/app-development.js" asp-append-version="true" type="text/javascript"></script>
+<script src="http://localhost:8080/1.app-development.js" asp-append-version="true" type="text/javascript"></script>
+<script src="http://localhost:8080/2.app-development.js" asp-append-version="true" type="text/javascript"></script>
+<script src="http://localhost:8080/3.app-development.js" asp-append-version="true" type="text/javascript"></script>
+<script src="http://localhost:8080/4.app-development.js" asp-append-version="true" type="text/javascript"></script>
