+++ src/studio/src/designer/frontend/package.json
@@ -27,6 +27,7 @@
     "app-development",
     "shared",
     "ux-editor",
-    "packages/schema-editor"
+    "packages/schema-editor",
+    "packages/schema-model"
   ]
 }
