+++ src/studio/src/designer/frontend/packages/schema-editor/package.json
@@ -9,6 +9,7 @@
   "types": "dist/index.d.ts",
   "peerDependencies": {
     "@altinn/altinn-design-system": "^0.7.1 || ^0.8.0",
+    "@altinn/schema-model": "*",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
     "@material-ui/lab": "^4.0.0-alpha.61",
@@ -74,7 +75,8 @@
     "moduleNameMapper": {
       "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/test/__mocks__/fileMock.js",
       "\\.(css|less)$": "<rootDir>/test/__mocks__/styleMock.js",
-      "^app-shared/(.*)": "<rootDir>/../../shared/$1"
+      "^app-shared/(.*)": "<rootDir>/../../shared/$1",
+      "^@altinn/schema-model/(.*)": "<rootDir>/../schema-model/$1"
     },
     "testRegex": "(test/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
     "moduleFileExtensions": [
