--- src/studio/src/designer/frontend/packages/schema-editor/src/ui-schema/index.ts
@@ -1,9 +1,0 @@
-import { UiSchemaMap } from './types';
-import { createPointerLookupTable } from './utils';
-
-export * from './build-ui-schema';
-export * from './build-json-schema';
-
-export const isPointerInUse = (uiNodeMap: UiSchemaMap, pointer: string) => {
-  return createPointerLookupTable(uiNodeMap).has(pointer);
-};
