--- src/studio/src/designer/frontend/packages/schema-editor/src/ui-schema/prettier.config.js
@@ -1,5 +1,0 @@
-const projectConfig = require('../../../../../../../../../prettier.config');
-module.exports = {
-  ...projectConfig,
-  printWidth: 100,
-};
