+++ src/studio/src/designer/frontend/packages/schema-editor/src/utils/ui-schema-utils.ts
@@ -1,5 +1,5 @@
 import { UiSchemaItem } from '../types';
-import { ObjectKind } from '../types/enums';
+import { ObjectKind } from '@altinn/schema-model';
 
 export function getObjectKind(item?: UiSchemaItem): ObjectKind {
   if (item?.$ref !== undefined || item?.items?.$ref !== undefined) {
