+++ src/studio/src/designer/frontend/packages/schema-model/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "@altinn/schema-model",
+  "version": "0.1.0",
+  "private": true,
+  "description": "JSON schema internal model",
+  "main": "src/index.ts",
+  "peerDependencies": {
+    "jsonpointer": "^4.1.0 || ^5.0.0"
+  },
+  "devDependencies": {
+    "jest": "^29.0.3"
+  },
+  "scripts": {
+    "build": "cross-env NODE_ENV=production webpack --config webpack.config.production.js --progress",
+    "test": "jest"
+  },
+  "jest": {
+    "transform": {
+      ".(ts|tsx)": "ts-jest"
+    },
+    "reporters": [
+      "default",
+      "jest-junit"
+    ],
+    "testRegex": "(test/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
+    "moduleFileExtensions": [
+      "ts",
+      "js"
+    ],
+    "testEnvironment": "jsdom"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
+}
