+++ src/studio/src/designer/frontend/packages/schema-model/prettier.config.js
@@ -1,0 +1,5 @@
+const projectConfig = require('../../../../../../../prettier.config');
+module.exports = {
+  ...projectConfig,
+  printWidth: 100,
+};
