+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/handlers/generic.ts
