+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/handlers/restrictions.ts
@@ -1,5 +1,5 @@
-import { AllRestrictions } from '../../utils/restrictions';
 import { JsonSchemaNode } from '../types';
+import { AllRestrictions } from '../restrictions';
 
 /**
  * Handling restrictions
