+++ src/studio/src/designer/frontend/packages/schema-model/src/setupTests.ts
@@ -1,0 +1,1 @@
+import 'jest';
